requirements.txt
pick==2.2.0
pyperclip==1.8.2
PyYAML==6.0